% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_funcs.R
\name{err_fun}
\alias{err_fun}
\alias{poi_dev}
\alias{mse}
\alias{wgt_mse}
\title{Predefined error functions}
\usage{
err_fun(y_true, y_pred, w_case)

poi_dev(y_true, y_pred)

mse(y_true, y_pred)

wgt_mse(y_true, y_pred, w_case)
}
\arguments{
\item{y_true}{vector of the true target values.}

\item{y_pred}{vector of the predicted target values.}

\item{w_case}{vector of case weights.}
}
\description{
Some predefined error functions to use in \code{\link{autotune}}.
}
\section{Functions}{
\itemize{
\item \code{poi_dev}: Poisson deviance loss function

\item \code{mse}: Mean squared error loss function

\item \code{wgt_mse}: Weighted mean squared error loss function
}}

\keyword{internal}
