% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain predictions}
\usage{
explain(surro, instance, plt = TRUE)
}
\arguments{
\item{surro}{The surrogate GLM fit (i.e., a "glm" object).}

\item{instance}{A single row data frame with the instance to be explained.}

\item{plt}{Boolean whether to return a ggplot or the underlying data.}
}
\value{
A tidy data frame or ggplot with each feature's contribution to the
prediction of model \code{surro} on observation \code{instance}.
}
\description{
Explain the surrogate GLM via each feature's contribution to the prediction.
}
\examples{
\dontrun{
data('mtpl_be')
features <- setdiff(names(mtpl_be),c('id', 'nclaims', 'expo'))
set.seed(12345)
gbm_fit <- gbm::gbm(as.formula(paste('nclaims ~',
                               paste(features, sep = ' ', collapse = ' + '))),
                    distribution = 'poisson',
                    data = mtpl_be,
                    n.trees = 50,
                    interaction.depth = 3,
                    shrinkage = 0.1)
gbm_fun <- function(object, newdata) mean(predict(object, newdata, n.trees = object$n.trees, type = 'response'))
data_segm <- gbm_fit \%>\% insights(vars = c('ageph', 'bm', 'coverage', 'fuel'),
                                  data = mtpl_be,
                                  interactions = 'auto',
                                  hcut = 0.7,
                                  pred_fun = gbm_fun) \%>\%
                         segmentation(data = mtpl_be,
                                      lambda = 0.001)
data_segm \%>\% surrogate(formula = nclaims ~ ageph_ + bm_ + fuel_ + coverage_ + ageph_bm_ + bm_fuel_,
                        family =  poisson(link = 'log'),
                        offset = log(expo)) \%>\%
              explain(instance = data_segm[34, ])
}
}
