% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vi.R
\name{get_vi}
\alias{get_vi}
\title{Get the variable importance}
\usage{
get_vi(mfit, ...)
}
\arguments{
\item{mfit}{A fitted model object (e.g., a "gbm" or "randomForest" object).}

\item{...}{Additional optional arguments to be passed onto
\code{\link[vip:vi]{vip::vi}}.}
}
\value{
A tidy data frame (i.e., a "tibble" object) with two columns:
\code{Variable} and \code{Importance}.
}
\description{
Compute variable importance scores for the predictors in a model. Note that
\code{get_vi} is a simple wrapper for \code{\link[vip:vi]{vip::vi}}.
}
\examples{
\dontrun{
data('mtpl_be')
features <- setdiff(names(mtpl_be),c('id', 'nclaims', 'expo'))
set.seed(12345)
gbm_fit <- gbm::gbm(as.formula(paste('nclaims ~',
                               paste(features, sep = ' ', collapse = ' + '))),
                    distribution = 'poisson',
                    data = mtpl_be,
                    n.trees = 50,
                    interaction.depth = 3,
                    shrinkage = 0.1)
gbm_fit \%>\% get_vi
gbm_fit \%>\% get_vi(scale = TRUE)
}
}
\seealso{
\code{\link[vip]{vi}} and
\url{https://CRAN.R-project.org/package=vip}.
}
