% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_pd.R
\name{group_pd}
\alias{group_pd}
\alias{group_pd_ckseg}
\alias{group_pd_ckmns}
\title{Partial dependence grouping}
\usage{
group_pd(pd, ngroups)

group_pd_ckseg(pd, ngroups)

group_pd_ckmns(pd, ngroups)
}
\arguments{
\item{pd}{A data frame containing the partial dependence function as returned
by \code{\link{get_pd}}.}

\item{ngroups}{An integer specifying the number of groups.}
}
\value{
The tidy data frame (i.e., a "tibble" object) supplied in \code{pd}
with three additional columns: xgrp, ygrp and wgrp. Column \code{xgrp}
contains feature groups, column \code{ygrp} the average partial dependence
for the group and \code{wgrp} the sum of observation counts for the group.
}
\description{
Grouping of feature values/levels by binning continuous/ordinal features and
clustering nominal features. Partial dependencies are used to perform the
grouping in a data-driven way.
}
\section{Functions}{
\itemize{
\item \code{group_pd_ckseg}: Grouping via \code{\link[Ckmeans.1d.dp]{Cksegs.1d.dp}}.

\item \code{group_pd_ckmns}: Grouping via \code{\link[Ckmeans.1d.dp]{Ckmeans.1d.dp}}.
}}

\examples{
\dontrun{
data('mtpl_be')
features <- setdiff(names(mtpl_be),c('id', 'nclaims', 'expo'))
set.seed(12345)
gbm_fit <- gbm::gbm(as.formula(paste('nclaims ~',
                               paste(features, sep = ' ', collapse = ' + '))),
                    distribution = 'poisson',
                    data = mtpl_be,
                    n.trees = 50,
                    interaction.depth = 3,
                    shrinkage = 0.1)
gbm_fun <- function(object, newdata) mean(predict(object, newdata, n.trees = object$n.trees, type = 'response'))
gbm_fit \%>\% get_pd(var = 'ageph',
                   grid = data.frame('ageph' = 20:90),
                   data = mtpl_be,
                   subsample = 10000,
                   fun = gbm_fun) \%>\%
            group_pd(ngroups = 5)
}
}
