% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_pd.R
\name{interaction_pd}
\alias{interaction_pd}
\title{Pure two-way interaction}
\usage{
interaction_pd(pd_2d, pd_1d)
}
\arguments{
\item{pd_2d}{A data frame containing the 2D partial dependence as returned by
\code{\link{get_pd}} with \code{var = 'var1_var2'}.}

\item{pd_1d}{A list of data frames containing the 1D partial dependence for
var1 and var2 as returned by \code{\link{get_pd}} with \code{var = 'var1'}
and \code{var = 'var2'} respectively.}
}
\value{
A tidy data frame (i.e., a "tibble" object) with four (x1, x2, y, w)
columns. Columns \code{x} contain variable values, column \code{y} the
partial dependence and \code{w} the observation counts.
}
\description{
Compute the pure two-way interaction partial dependence function for two
features from the 2D partial dependence and both 1D partial dependencies.
}
\examples{
\dontrun{
data('mtpl_be')
features <- setdiff(names(mtpl_be),c('id', 'nclaims', 'expo'))
set.seed(12345)
gbm_fit <- gbm::gbm(as.formula(paste('nclaims ~',
                               paste(features, sep = ' ', collapse = ' + '))),
                    distribution = 'poisson',
                    data = mtpl_be,
                    n.trees = 50,
                    interaction.depth = 3,
                    shrinkage = 0.1)
gbm_fun <- function(object, newdata) mean(predict(object, newdata, n.trees = object$n.trees, type = 'response'))
pd_2d <- get_pd(mfit = gbm_fit,
                var = 'ageph_power',
                grid = expand.grid('ageph' = 20:60, 'power' = 40:70),
                data = mtpl_be,
                subsample = 10000
                fun = gbm_fun)
pd_1d <- list(get_pd(mfit = gbm_fit,
                     var = 'ageph',
                     grid = data.frame('ageph' = 20:60),
                     data = mtpl_be,
                     subsample = 10000
                     fun = gbm_fun),
              get_pd(mfit = gbm_fit,
                     var = 'power',
                     grid = data.frame('power' = 40:70),
                     data = mtpl_be,
                     subsample = 10000
                     fun = gbm_fun))
interaction_pd(pd_2d, pd_1d)
}
}
