% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_grid.R
\name{lambda_grid}
\alias{lambda_grid}
\title{Lambda grid}
\usage{
lambda_grid(
  fx_vars,
  lambda_range = as.vector(outer(seq(1, 10, 0.1), 10^(-7:3)))
)
}
\arguments{
\item{fx_vars}{A list of data frames containing the feature effects.}

\item{lambda_range}{A numeric vector of possible values for lambda.}
}
\value{
A tidy data frame (i.e., a "tibble" object) with the lambda grid.
}
\description{
Create a grid of lambda values that result in a unique grouping. Main and
interaction effects are dealt with seperately due to the scale difference.
}
\examples{
\dontrun{
data('mtpl_be')
features <- setdiff(names(mtpl_be),c('id', 'nclaims', 'expo'))
set.seed(12345)
gbm_fit <- gbm::gbm(as.formula(paste('nclaims ~',
                               paste(features, sep = ' ', collapse = ' + '))),
                    distribution = 'poisson',
                    data = mtpl_be,
                    n.trees = 50,
                    interaction.depth = 3,
                    shrinkage = 0.1)
gbm_fun <- function(object, newdata) mean(predict(object, newdata, n.trees = object$n.trees, type = 'response'))
gbm_fit \%>\% insights(vars = c('ageph', 'bm', 'coverage', 'fuel'),
                     data = mtpl_be,
                     interactions = 'auto',
                     hcut = 0.7,
                     pred_fun = gbm_fun) \%>\%
            lambda_grid
}
}
