% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maidrr-package.R
\docType{package}
\name{maidrr-package}
\alias{maidrr-package}
\title{maidrr: Model-Agnostic Interpretable Data-driven suRRogate}
\description{
The goal of maidrr is to aid you in the development of a Model-Agnostic Interpretable Data-driven suRRogate for your black box algorithm of choice.
In short, these are the steps in the procedure:
\enumerate{
\item Partial dependencies (PDs) are used to obtain model insights from the black box in the form of feature effects.
\item Those effects are used to group values/levels within a feature in an optimal data-driven way, while performing built-in feature selection.
\item An interpretable GLM surrogate is fit to the segmented features. Meaningful interactions can be included if desired.
}
}
\keyword{package}
