% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtpl_fr}
\alias{mtpl_fr}
\title{French motor third party liability portfolio}
\format{A data frame with 163212 rows and 14 variables: \describe{
\item{id}{policyholder id}
\item{nclaims}{number of claims}
\item{expo}{exposure period, as a fraction of a year}
\item{fuel}{fuel type: diesel or gasoline}
\item{brand}{brand of the vehicle, 11 levels}
\item{region}{region of residence, 22 levels}
\item{power}{horsepower of the vehicle, 12 ordered levels}
\item{ageph}{age of the policyholder, in years}
\item{bm}{bonus-malus level, higher is worse}
\item{agec}{age of the vehicle, in years}
\item{popdens}{population density of the municipality of
residence, in persons per squared kilometer}
}}
\usage{
mtpl_fr
}
\description{
A dataset containing the number of motor insurance claims and several
characteristics for a portfolio of around 669000 policyholders.
}
\keyword{datasets}
