% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pd.R
\name{plot_pd}
\alias{plot_pd}
\title{Plot partial dependence}
\usage{
plot_pd(pd)
}
\arguments{
\item{pd}{Data frame containing the partial dependence effect as returned by
\code{\link{get_pd}}.}
}
\value{
ggplot object showing the partial dependence effect in \code{pd}.
}
\description{
Plot the partial dependence functions (i.e., marginal effects) for the
predictors in a model.
}
\examples{
\dontrun{
data('mtpl_be')
features <- setdiff(names(mtpl_be), c('id', 'nclaims', 'expo', 'long', 'lat'))
set.seed(12345)
gbm_fit <- gbm::gbm(as.formula(paste('nclaims ~',
                               paste(features, collapse = ' + '))),
                    distribution = 'poisson',
                    data = mtpl_be,
                    n.trees = 50,
                    interaction.depth = 3,
                    shrinkage = 0.1)
gbm_fun <- function(object, newdata) mean(predict(object, newdata, n.trees = object$n.trees, type = 'response'))
gbm_fit \%>\% get_pd(var = 'ageph',
                   grid = 'ageph' \%>\% get_grid(data = mtpl_be),
                   data = mtpl_be,
                   subsample = 10000,
                   fun = gbm_fun) \%>\%
            plot_pd
}
}
