% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_lvls}
\alias{rm_lvls}
\title{Drop unused factor levels}
\usage{
rm_lvls(fit, test_data)
}
\arguments{
\item{fit}{Model fit of class "glm".}

\item{test_data}{Data frame containing the test data.}
}
\description{
Function to drop empty factor levels in test data
(\href{https://stackoverflow.com/a/39495480/4185785}{source}).
}
