% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation}
\alias{segmentation}
\title{Segmentation of the data}
\usage{
segmentation(fx_vars, data, lambda)
}
\arguments{
\item{fx_vars}{A list of data frames containing the feature effects.}

\item{data}{Data frame containing the original training data.}

\item{lambda}{A numeric (same lambda for all features in \code{fx_vars}) or a
numeric vector of \code{length(fx_vars)} (for feature-specific lambdas).}
}
\value{
A tidy data frame (i.e., a "tibble" object) with the segmented data.
The grouped features all have a trailing underscore in their name.
}
\description{
Segmentation of observations based on the grouping of feature effects.
}
\examples{
\dontrun{
data('mtpl_be')
features <- setdiff(names(mtpl_be),c('id', 'nclaims', 'expo'))
set.seed(12345)
gbm_fit <- gbm::gbm(as.formula(paste('nclaims ~',
                               paste(features, sep = ' ', collapse = ' + '))),
                    distribution = 'poisson',
                    data = mtpl_be,
                    n.trees = 50,
                    interaction.depth = 3,
                    shrinkage = 0.1)
gbm_fun <- function(object, newdata) mean(predict(object, newdata, n.trees = object$n.trees, type = 'response'))
gbm_fit \%>\% insights(vars = c('ageph', 'bm', 'coverage', 'fuel'),
                     data = mtpl_be,
                     interactions = 'auto',
                     hcut = 0.7,
                     pred_fun = gbm_fun) \%>\%
            segmentation(data = mtpl_be,
                         lambda = 0.01)
}
}
