% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.R
\name{get_grid}
\alias{get_grid}
\title{Get feature grid}
\usage{
get_grid(var, data)
}
\arguments{
\item{var}{Character string giving the name of the feature of interest. Only
integer, numeric or factor variables are supported for now.}

\item{data}{Data frame containing the original training data.}
}
\value{
One-column tidy data frame (i.e., a "tibble" object). This column
contains the grid values for feature \code{var} based on the observed
values in \code{data}.
}
\description{
Get the grid values for a feature based on the observed values in the data.
}
\examples{
\dontrun{
data('mtpl_be')
'ageph' \%>\% get_grid(data = mtpl_be)
'coverage' \%>\% get_grid(data = mtpl_be)
tidyr::expand_grid('ageph' \%>\% get_grid(data = mtpl_be),
                   'coverage' \%>\% get_grid(data = mtpl_be))
}
}
