% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maidrr-package.R
\docType{package}
\name{maidrr-package}
\alias{maidrr-package}
\title{maidrr: A package that implements mAIDs in R}
\description{
The maidrr package provides functions to aid you in developing a
model-Agnostic Interpretable Data-driven suRRogate model for
any type of black box algorithm.
}
\keyword{package}
